pyproject.toml
[project]
name = "repl-nix-ecom-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask==3.1.0",
    "psycopg2-binary==2.9.10",
]
